Bed fans are trurned of while heating the bed
because sometime the printer complains about the heating isnt working
in the desired rate. The printer isnt expecting the heating is so slow.

It takes allmost the same time to heat up the bed turn mon the bed fans
and start the print after the chamber temperrature hits the desired
temperature to start the print.